chore: Renamed test-app-for-sample-v2-addon to test-app-for-my-v2-addon
