+++ .changeset/angry-cars-raise.md
@@ -6,7 +6,7 @@
 "test-app-for-embroider-css-modules": patch
 "my-app": patch
 "type-css-modules": patch
-"test-app-for-sample-v2-addon": patch
+"test-app-for-my-v2-addon": patch
 "my-v2-addon": patch
 ---
