+++ tests/my-v2-addon/.editorconfig
