+++ tests/my-v2-addon/.eslintrc.js
