+++ tests/my-v2-addon/.gitignore
