+++ tests/my-v2-addon/.prettierrc.js
