+++ tests/my-v2-addon/.stylelintignore
