+++ tests/my-v2-addon/.stylelintrc.js
