+++ tests/my-v2-addon/.template-lintrc.js
