+++ tests/my-v2-addon/.watchmanconfig
