+++ tests/my-v2-addon/app/config/environment.d.ts
