+++ tests/my-v2-addon/app/index.html
@@ -12,7 +12,7 @@
     <link rel="apple-touch-icon" type="image/png" sizes="256x256" href="/assets/favicon.png">
     <link href="https://fonts.googleapis.com/css?family=Raleway:400,700&display=swap" rel="stylesheet">
     <link integrity="" rel="stylesheet" href="{{rootURL}}assets/vendor.css">
-    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/test-app-for-sample-v2-addon.css">
+    <link integrity="" rel="stylesheet" href="{{rootURL}}assets/test-app-for-my-v2-addon.css">
 
     {{content-for "head-footer"}}
   </head>
@@ -20,7 +20,7 @@
     {{content-for "body"}}
 
     <script src="{{rootURL}}assets/vendor.js"></script>
-    <script src="{{rootURL}}assets/test-app-for-sample-v2-addon.js"></script>
+    <script src="{{rootURL}}assets/test-app-for-my-v2-addon.js"></script>
 
     {{content-for "body-footer"}}
   </body>
