+++ tests/my-v2-addon/app/router.ts
