+++ tests/my-v2-addon/app/templates/application.hbs
