+++ tests/my-v2-addon/app/templates/index.hbs
