+++ tests/my-v2-addon/config/ember-cli-update.json
