+++ tests/my-v2-addon/config/ember-try.js
