+++ tests/my-v2-addon/config/optional-features.json
