+++ tests/my-v2-addon/config/targets.js
