+++ tests/my-v2-addon/ember-cli-build.js
