+++ tests/my-v2-addon/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "test-app-for-sample-v2-addon",
+  "name": "test-app-for-my-v2-addon",
   "version": "1.0.0",
   "private": true,
-  "description": "test-app for my-v2-addon",
+  "description": "Test app for my-v2-addon",
   "keywords": [
     "css",
     "css-modules",
