+++ tests/my-v2-addon/public/assets/favicon.png
