+++ tests/my-v2-addon/public/robots.txt
