+++ tests/my-v2-addon/tests/index.html
@@ -11,7 +11,7 @@
 
     <link href="https://fonts.googleapis.com/css?family=Raleway:400,700&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="{{rootURL}}assets/vendor.css">
-    <link rel="stylesheet" href="{{rootURL}}assets/test-app-for-sample-v2-addon.css">
+    <link rel="stylesheet" href="{{rootURL}}assets/test-app-for-my-v2-addon.css">
     <link rel="stylesheet" href="{{rootURL}}assets/test-support.css">
 
     {{content-for "head-footer"}}
@@ -31,7 +31,7 @@
     <script src="/testem.js" integrity="" data-embroider-ignore></script>
     <script src="{{rootURL}}assets/vendor.js"></script>
     <script src="{{rootURL}}assets/test-support.js"></script>
-    <script src="{{rootURL}}assets/test-app-for-sample-v2-addon.js"></script>
+    <script src="{{rootURL}}assets/test-app-for-my-v2-addon.js"></script>
     <script src="{{rootURL}}assets/tests.js"></script>
 
     {{content-for "body-footer"}}
