+++ tests/my-v2-addon/tests/integration/components/ui/page-test.ts
