+++ tests/my-v2-addon/tests/test-helper.ts
@@ -2,8 +2,8 @@ import { setApplication } from '@ember/test-helpers';
 import { start } from 'ember-qunit';
 import * as QUnit from 'qunit';
 import { setup } from 'qunit-dom';
-import Application from 'test-app-for-sample-v2-addon/app';
-import config from 'test-app-for-sample-v2-addon/config/environment';
+import Application from 'test-app-for-my-v2-addon/app';
+import config from 'test-app-for-my-v2-addon/config/environment';
 
 setApplication(Application.create(config.APP));
