+++ tests/my-v2-addon/types/global.d.ts
@@ -1,5 +1,5 @@
 // Types for compiled templates
-declare module 'test-app-for-sample-v2-addon/templates/*' {
+declare module 'test-app-for-my-v2-addon/templates/*' {
   import { TemplateFactory } from 'ember-cli-htmlbars';
 
   const tmpl: TemplateFactory;
