+++ tests/my-v2-addon/types/test-app-for-my-v2-addon/index.d.ts
